requirements.txt
Flask==2.0.2
google-cloud-storage==1.42.3
Werkzeug==2.0.2

